+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+## TBD
+
+* Amend secondary instance URL to bugsnag.smartbear.com []()
+
 ## 4.1.0 (2025-07-17)
 
 ### Enhancements